refactor: generalization of structures
